refactor: quest select fields 관련 constant 생성
