+++ server/src/common/constants/quest.constant.ts
@@ -1,0 +1,12 @@
+export const QUEST_SELECT_FIELDS = [
+  'id',
+  'title',
+  'difficulty',
+  'mode',
+  'startDate',
+  'endDate',
+  'hidden',
+  'status',
+  'createdAt',
+  'updatedAt',
+];
