+++ server/src/common/constants/transaction.constant.ts
