+++ server/src/core/middlewares/transaction.middleware.ts
@@ -2,7 +2,7 @@ import { Injectable, NestMiddleware } from '@nestjs/common';
 import { Namespace, createNamespace, getNamespace } from 'cls-hooked';
 import { NextFunction, Request, Response } from 'express';
 import { EntityManager } from 'typeorm';
-import { ENTITY_MANAGER, TRANSACTION } from '../../common/utils/constants.util';
+import { ENTITY_MANAGER, TRANSACTION } from '../../common/constants/transaction.constant';
 
 @Injectable()
 export class TransactionMiddleware implements NestMiddleware {
